requirements.txt
flask
pytest
python-dotenv
requests
loguru